backend/requirements.txt
django==5.2.5
asgiref
djangorestframework==3.16.1
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
pytz==2025.2
sqlparse
psycopg2-binary==2.9.10
python-dotenv==1.1.1
Pillow==11.0.0
